Move epwreader to folder utils
Organising a bit demand model. nothing in the structure was changed
